Add name change checking and don't worry about recent data
We don't really care if, when the user updates their settings, that the player that they have in their account has recent data within 24 hours. That could potentially ruin the image that is being produced. All that we care about is that there is at least *some* data that we can compare against.